just_run.js; TODO: support deck sizes other than eight
